factor detector

The factor detector q-statistic measures the spatial stratified heterogeneity of a variable Y, or the determinant power of a covariate X of Y.

Usage

factor_detector(y, x)

Arguments

y

Variable Y, continuous numeric vector.

x

Covariable X, factor, character or discrete numeric.

Value

A list contains the Q-statistic and the P-value.

Author

Wenbo Lv lyu.geosocial@gmail.com

Examples

factor_detector(y = 1:7,x = c('x',rep('y',3),rep('z',3)))
#> $`Q-statistic`
#> [1] 0.7714286
#> 
#> $`P-value`
#> [1] 0.07936477
#>