interactive detector for spatial associations(IDSA)

Function for interactive detector for spatial associations model.

Usage

idsa(
  formula,
  data,
  wt = NULL,
  overlaymethod = "and",
  locations = NULL,
  discvar = NULL,
  discnum = NULL,
  discmethod = NULL,
  strategy = 2L,
  increase_rate = 0.05,
  cores = 6,
  seed = 123456789,
  alpha = 0.95,
  ...
)

Arguments

formula

A formula of IDSA model.

data

A data.frame or tibble of observation data.

wt

(optional) The spatial weight matrix.When wt is not provided, must provide locations. And gdverse will use locations columns to construct spatial weight use inverse_distance_weight().

overlaymethod

(optional) Spatial overlay method. One of and, or, intersection. Default is and.

locations

(optional) The geospatial locations coordinate columns name which in data. Useful and must provided when wt is not provided. When wt is provided, locations is not need.

discvar

Name of continuous variable columns that need to be discretized.Noted that when formula has discvar, data must have these columns.

discnum

(optional) Number of multilevel discretization.Default will use 3:22.

discmethod

(optional) The discretization methods. Default all use quantile. More details to see st_unidisc().

strategy

(optional) Discretization strategy. When strategy is 1L, choose the highest SPADE model q-statistics to determinate optimal spatial data discretization parameters. When strategy is 2L, The optimal discrete parameters of spatial data are selected by combining LOESS model.

increase_rate

(optional) The critical increase rate of the number of discretization. Default is 5%.

cores

(optional) A positive integer(default is 1). If cores > 1, a 'parallel' package cluster with that many cores is created and used. You can also supply a cluster object.

seed

(optional) Random number seed, default is 123456789.

alpha

(optional) Specifies the size of confidence level.Default is 0.95.

...

(optional) Other arguments passed to cpsd_disc().

Value

A list with PID values tibble under different spatial overlays and performance evaluation indicators.

Note

The IDSA model requires at least \(2^n-1\) calculations when has \(n\) explanatory variables. When there are more than 10 explanatory variables, carefully consider the computational burden of this model. When there are a large number of explanatory variables, the data dimensionality reduction method can be used to ensure the trade-off between analysis results and calculation speed.

References

Yongze Song & Peng Wu (2021) An interactive detector for spatial associations, International Journal of Geographical Information Science, 35:8, 1676-1701, DOI:10.1080/13658816.2021.1882680

Author

Wenbo Lv lyu.geosocial@gmail.com

Examples

if (FALSE) {
data('sim')
g = idsa(y ~ ., data = sim, locations = c('lo','la'),
         discvar = c("xa","xb","xc"), cores = 6)
g
}