robust geographical detector(RGD) model

Function for robust geographical detector(RGD) model.

Usage

rgd(
  formula,
  data,
  discvar,
  discnum = NULL,
  minsize = NULL,
  cores = 1,
  type = "factor",
  alpha = 0.95,
  ...
)

Arguments

formula

A formula of RGD model.

data

A data.frame or tibble of observation data.

discvar

Name of continuous variable columns that need to be discretized.Noted that when formula has discvar, data must have these columns.

discnum

A numeric vector of discretized classes of columns that need to be discretized. Default all discvar use 10.

minsize

(optional) The min size of each discretization group.Default all use 1.

cores

positive integer(default is 1). If cores > 1, use python's joblib package to parallel computation.

type

(optional) The type of geographical detector,which must be factor(default), interaction, risk, ecological.You can run one or more at a time.

alpha

(optional) Specifies the size of confidence level.Default is 0.95.

...

(optional) Other arguments passed to robust_disc().

Value

A list of the RGD model result.

References

Zhang, Z., Song, Y.*, & Wu, P., 2022. Robust geographical detector. International Journal of Applied Earth Observation and Geoinformation. 109, 102782. DOI: 10.1016/j.jag.2022.102782.

Author

Wenbo Lv lyu.geosocial@gmail.com

Examples

if (FALSE) {
library(terra)
library(tidyverse)
fvcpath = "https://github.com/SpatLyu/rdevdata/raw/main/FVC.tif"
fvc = terra::rast(paste0("/vsicurl/",fvcpath))
fvc = terra::aggregate(fvc,fact = 5)
fvc = as_tibble(terra::as.data.frame(fvc,na.rm = T))
rgd(fvc ~ ., data = fvc, discnum = 10,
    discvar = names(select(fvc,-c(fvc,lulc))),
    cores = 6, type =c('factor','interaction'))
}