power of spatial and multilevel discretization determinant(PSMD) and the corresponding pseudo-p value

Function for calculate power of spatial and multilevel discretization determinant and the corresponding pseudo-p value.

Usage

psmd_pseudop(
  formula,
  data,
  wt = NULL,
  locations = NULL,
  discnum = NULL,
  discmethod = NULL,
  cores = 6,
  seed = 123456789,
  permutations = 99,
  ...
)

Arguments

formula

A formula of calculate power of spatial and multilevel discretization determinant \(PSMDQ_s\).

data

A data.frame or tibble of observation data.

wt

(optional) The spatial weight matrix.When wt is not provided, must provide locations. And gdverse will use locations columns to construct spatial weight use inverse_distance_weight().

locations

(optional) The geospatial locations coordinate columns name which in data. Useful and must provided when wt is not provided.

discnum

(optional) Number of multilevel discretization.Default will use 3:15.

discmethod

(optional) The discretization methods. Default will use quantile. When discmethod is robust use robust_disc(), others use st_unidisc().Now only support one discmethod at one time.

cores

(optional) A positive integer(default is 1). If cores > 1, use parallel computation.

seed

(optional) Random seed number, default is 123456789.

permutations

(optional) The number of permutations for the PSD computation. Default is 99.

...

(optional) Other arguments passed to st_unidisc() or robust_disc().

Value

A list of power of spatial and multilevel discretization determinant and the corresponding pseudo-p value.

Details

The power of spatial and multilevel discretization determinant formula is \(PSMDQ_s = MEAN(Q_s)\)

References

Xuezhi Cang & Wei Luo (2018) Spatial association detector (SPADE),International Journal of Geographical Information Science, 32:10, 2055-2075, DOI: 10.1080/13658816.2018.1476693

Author

Wenbo Lv lyu.geosocial@gmail.com

Examples

if (FALSE) {
library(sf)
usfi = read_sf(system.file('extdata/USFI_Xian.gpkg',package = 'gdverse')) |>
  dplyr::select(dplyr::all_of(c("NDVI","BH","SUHI")))
coord = usfi |>
  st_centroid() |>
  st_coordinates()
usfi = usfi |>
  dplyr::bind_cols(coord) |>
  st_drop_geometry()
tictoc::tic()
psmd_pseudop('SUHI ~ BH',data = dplyr::select(usfi,SUHI,BH,X,Y),
             locations = c('X','Y'),cores = 6)
tictoc::toc()
}